+++ packages/orq-rc/src/orq_ai_sdk/models/createchunkop.py
@@ -257,7 +257,7 @@ class CreateChunkResponseBody(BaseModel):
     updated_by_id: str
     r"""The unique identifier of the user who updated the chunk"""
 
-    id: Optional[str] = "chunk_01JQXQEBBB9BB8J604ZG6RANEA"
+    id: Optional[str] = "chunk_01JQXQTYP7W7XG2C904M7ZY3P6"
     r"""Unique identifier for the element"""
 
     enabled: Optional[bool] = True