+++ packages/orq-rc/src/orq_ai_sdk/models/createcontactop.py
@@ -134,7 +134,7 @@ class CreateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-04-03T11:44:00.596Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-04-03T11:50:53.752Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")