+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JQXQEBB75ZQ1QSC3GP9KDFRV"
+        "01JQXQTYP3Z6FV35HMJKZ1SNDY"
     )
     r"""The id of the resource"""