+++ packages/orq-rc/RELEASES.md
@@ -1149,3 +1149,13 @@ Based on:
 - [python v3.3.0-rc.33] packages/orq-rc
 ### Releases
 - [PyPI v3.3.0-rc.33] https://pypi.org/project/orq-ai-sdk/3.3.0-rc.33 - packages/orq-rc
+
+## 2025-04-03 11:52:22
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.528.1 (2.565.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.3.0-rc.34] packages/orq-rc
+### Releases
+- [PyPI v3.3.0-rc.34] https://pypi.org/project/orq-ai-sdk/3.3.0-rc.34 - packages/orq-rc